move screen props definition to UI category
