--- Framework/Index/PropTypes.ts
@@ -1,4 +1,0 @@
-/**
- * Exposed to end users, export prop interfaces and types here.
- */
-export { IScreenRouteProps, RootStackPropsList, IScreenProps } from '../PropTypes/ScreenPropTypes';
