+++ projects/TemplateApp/Screen/SampleEmptyScreen.tsx
@@ -2,12 +2,11 @@ import React, { useContext, useState, useEffect, useCallback, useRef, memo } fro
 import * as UI from '../../../Framework/Index/UI';
 import * as Hook from '../../../Framework/Index/Hook';
 import * as Const from '../../../Framework/Index/Const';
-import * as PropTypes from '../../../Framework/Index/PropTypes';
 
 /**
  * Sample empty screen, duplicate for quick start for new screens.
  */
-const SampleEmptyScreen: React.FC<PropTypes.IScreenProps> = ({ navigation, route }) => {
+const SampleEmptyScreen: React.FC<UI.IScreenProps> = ({ navigation, route }) => {
   const { paramText } = route.params;
 
   return (
