+++ projects/TemplateApp/Screen/SampleTextInputBtnScreen.tsx
@@ -3,12 +3,11 @@ import { View } from 'react-native';
 import * as UI from '../../../Framework/Index/UI';
 import * as Hook from '../../../Framework/Index/Hook';
 import * as Const from '../../../Framework/Index/Const';
-import * as PropTypes from '../../../Framework/Index/PropTypes';
 
 /**
  * Showcase text, input and buttons.
  */
-const SampleTextInputBtnScreen: React.FC<PropTypes.IScreenProps> = ({ navigation, route }) => {
+const SampleTextInputBtnScreen: React.FC<UI.IScreenProps> = ({ navigation, route }) => {
   const [searchQuery, setSearchQuery] = useState('');
 
   return (
